pyproject.toml
[project]
name = "procgen-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pygame-ce>=2.5.6",
]
